+++ internal/controllers/control/cassandratask_controller_test.go
@@ -264,10 +264,10 @@ var _ = Describe("CassandraTask controller tests", func() {
 					Expect(len(completedTask.Status.Conditions)).To(Equal(2))
 					for _, cond := range completedTask.Status.Conditions {
 						switch cond.Type {
-						case api.JobComplete:
-							Expect(cond.Status).To(Equal(corev1.ConditionTrue))
-						case api.JobRunning:
-							Expect(cond.Status).To(Equal(corev1.ConditionFalse))
+						case string(api.JobComplete):
+							Expect(cond.Status).To(Equal(metav1.ConditionTrue))
+						case string(api.JobRunning):
+							Expect(cond.Status).To(Equal(metav1.ConditionFalse))
 						}
 					}
 				})
@@ -432,6 +432,20 @@ var _ = Describe("CassandraTask controller tests", func() {
 
 					Expect(completedTask.Status.Failed).To(BeNumerically(">=", nodeCount))
 				})
+				It("Replace a node in the datacenter without specifying the pod", func() {
+					testFailedNamespaceName := fmt.Sprintf("test-task-failed-%d", rand.Int31())
+					By("creating a datacenter", createDatacenter("dc1", testFailedNamespaceName))
+					By("Creating a task for replacenode")
+					taskKey, task := buildTask(api.CommandReplaceNode, testFailedNamespaceName)
+
+					Expect(k8sClient.Create(context.TODO(), task)).Should(Succeed())
+
+					completedTask := waitForTaskCompletion(taskKey)
+
+					Expect(completedTask.Status.Failed).To(BeNumerically(">=", 1))
+					Expect(completedTask.Status.Conditions[2].Type).To(Equal(string(api.JobFailed)))
+					Expect(completedTask.Status.Conditions[2].Message).To(Equal("valid pod_name to replace is required"))
+				})
 			})
 		})
 		Context("Sync jobs", func() {
@@ -468,7 +482,6 @@ var _ = Describe("CassandraTask controller tests", func() {
 
 			It("Runs a upgradesstables task against the datacenter pods", func() {
 				By("Creating a task for upgradesstables")
-				time.Sleep(1 * time.Second) // Otherwise the CreationTimestamp could be too new
 				taskKey := createTask(api.CommandUpgradeSSTables, testNamespaceName)
 
 				completedTask := waitForTaskCompletion(taskKey)
@@ -483,7 +496,6 @@ var _ = Describe("CassandraTask controller tests", func() {
 
 			It("Replaces a node in the datacenter", func() {
 				By("Creating a task for replacenode")
-				time.Sleep(1 * time.Second) // Otherwise the CreationTimestamp could be too new
 				taskKey, task := buildTask(api.CommandReplaceNode, testNamespaceName)
 
 				podKey := types.NamespacedName{
@@ -512,7 +524,6 @@ var _ = Describe("CassandraTask controller tests", func() {
 
 			It("Runs a flush task against the datacenter pods", func() {
 				By("Creating a task for flush")
-				time.Sleep(1 * time.Second) // Otherwise the CreationTimestamp could be too new
 				taskKey, task := buildTask(api.CommandFlush, testNamespaceName)
 				task.Spec.Jobs[0].Arguments.KeyspaceName = "ks1"
 				Expect(k8sClient.Create(context.Background(), task)).Should(Succeed())
@@ -529,7 +540,6 @@ var _ = Describe("CassandraTask controller tests", func() {
 
 			It("Runs a garbagecollect task against the datacenter pods", func() {
 				By("Creating a task for garbagecollect")
-				time.Sleep(1 * time.Second) // Otherwise the CreationTimestamp could be too new
 				taskKey, task := buildTask(api.CommandGarbageCollect, testNamespaceName)
 				task.Spec.Jobs[0].Arguments.KeyspaceName = "ks1"
 				Expect(k8sClient.Create(context.Background(), task)).Should(Succeed())